+++ src/ui/preferences/main.rs
@@ -32,6 +32,8 @@ pub enum PreferencesAppMsg {
 
     UpdateLauncherState,
     RepairGame,
+    RemakePrefix,
+    SetTimeoutFix(bool),
 
     Toast {
         title: String,
@@ -133,6 +135,17 @@ impl SimpleAsyncComponent for PreferencesApp {
                 PREFERENCES_WINDOW.as_ref().unwrap_unchecked().close();
 
                 sender.output(Self::Output::RepairGame);
+            }
+
+            #[allow(unused_must_use)]
+            PreferencesAppMsg::RemakePrefix => unsafe {
+                PREFERENCES_WINDOW.as_ref().unwrap_unchecked().close();
+
+                sender.output(Self::Output::RemakePrefix);
+            }
+
+            PreferencesAppMsg::SetTimeoutFix(value) => {
+                self.enhancements.emit(EnhancementsAppMsg::SetTimeoutFix(value));
             }
 
             PreferencesAppMsg::Toast { title, description } => unsafe {