+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.3] - 2026-05-08
+
+### Added
+
+- Added "Remake prefix" button
+- Added "Import game" button
+- Added setting to enable Winewayland to make the game use Wayland instead of X11
+- Added a setting to enable "Timeout fix"
+- Added automatic detection and suggestion for the "Timeout fix" setting
+
+### Fixed
+
+- Fixed a panic when the launcher folder is a broken symlink, the launcher will exit with an
+  error message instead
+- Fixed search url extractor
+
+### Changed
+
+- Changed .version file format to be a textual file taht can be edited in a text editor.
+  Old format files will still work correctly and will be overwritten on game update.
+
 ## [1.6.2] - 2026-01-03
 
 ### Added
@@ -130,7 +151,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <br>
 
-[unreleased]: https://github.com/an-anime-team/sleepy-launcher/compare/1.6.2...next
+[unreleased]: https://github.com/an-anime-team/sleepy-launcher/compare/1.6.3...next
+[1.6.3]: https://github.com/an-anime-team/sleepy-launcher/compare/1.6.2...1.6.3
 [1.6.2]: https://github.com/an-anime-team/sleepy-launcher/compare/1.6.1...1.6.2
 [1.6.1]: https://github.com/an-anime-team/sleepy-launcher/compare/1.6.0...1.6.1
 [1.6.0]: https://github.com/an-anime-team/sleepy-launcher/compare/1.5.0...1.6.0